Show game played-time in admin activity & player profiles; drop admin Replays tab
Recent-games lists previously showed only the date (endedAt.slice(0,10)). Add the
time the game was played via a shared formatDateTime helper, applied to:
- admin Activity "Recent games"
- signed-in profile history (ProfilePage)
- public-profile / shared StatsDashboard recent games

Also remove the admin Replays section — the same games (with timestamps, modes,
players, and a Watch link) are now in Activity. This deletes the redundant admin
hub tile, the AdminPage replays route, and the dead AdminController
(/api/admin/games[/{id}/replay]); the bootstrap-password check now validates against
/api/admin/users. The shared ReplayViewer component (used by in-game replay) and the
GameSummary DTO (used by the player /api/replays controller) are kept — GameSummary
moves to PlayerReplayController next to its toSummary() extension.
