+++ game-server/src/main/kotlin/com/wingedsheep/gameserver/controller/AdminStatsController.kt
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.RestController
 
 /**
  * Admin-only global stats for the dashboard: totals, games-per-day, mode/color distributions, and an
- * IP-based geolocation estimate. Auth reuses the same `X-Admin-Password` header as [AdminController].
+ * IP-based geolocation estimate. Auth uses the shared `X-Admin-Password` / admin-account gate.
  * Mounted only when accounts are enabled (the stats live in Postgres); the geolocation endpoint
  * resolves raw IPs server-side and returns only aggregated locations — raw IPs never reach the client.
  */
